% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_fairness.R
\name{analysis_estimation}
\alias{analysis_estimation}
\title{Main function: Estimation of nuisance parameters and unfairness metrics.}
\usage{
analysis_estimation(
  data,
  cutoff,
  pi_model,
  pi_model_type,
  pi_model_seed = NULL,
  f_lasso = NULL,
  xvars,
  gen_null = F,
  R_null = 1000,
  bootstrap = "none",
  m_factor = 0.75
)
}
\arguments{
\item{data}{Dataframe or tibble. Must include at least A1, A2, Y (binary 0/1),
D (binary 0/1), covariates used to train propensity score model,
S_prob (continuous risk probability).}

\item{cutoff}{Classification cutoff. Must be a single numeric value.}

\item{pi_model}{Propensity score model object.}

\item{pi_model_type}{Type of propensity score model. Options are 'glm' (GLM).
or 'rf' (random forest)}

\item{pi_model_seed}{Random seed for random forest propensity score model.}

\item{f_lasso}{Formula for GLM propensity score model.}

\item{xvars}{Character vector of covariates used in propensity score model.}

\item{gen_null}{T/F: generate null distributions.}

\item{R_null}{Number of replications for null distribution.}

\item{bootstrap}{Obtain bootstrap estimates using rescaled method ('rescaled') or
no bootstrap ('none', the default).}

\item{m_factor}{Fractional power for calculating resample size (default 0.75).}
}
\value{
List with the following components:
\itemize{
\item defs: estimated definitions
\item estY0: mean estimated Y0 for each protected group
\item table_null: null distribution table (if gen_null = T)
\item boot_out: bootstrap estimates for metrics and error rates (if bootstrap = 'rescaled')
\item est.choice: Input data frame with Y0 estimates added
}
}
\description{
Main function: Estimation of nuisance parameters and unfairness metrics.
}
