% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_fairness.R
\name{analysis_nulldist}
\alias{analysis_nulldist}
\title{Simulate null distributions.}
\usage{
analysis_nulldist(
  data,
  R,
  cutoff,
  xvars,
  pi_model_type,
  pi_model_seed = NULL,
  f_lasso = NULL,
  defs_length_base
)
}
\arguments{
\item{data}{Dataframe or tibble. Must include at least A1, A2, Y (binary 0/1),
D (binary 0/1), covariates used to train propensity score model,
S_prob (continuous risk probability).}

\item{R}{Number of replications.}

\item{cutoff}{Classification cutoff. Must be a single numeric value.}

\item{xvars}{Character vector of covariates used in propensity score model.}

\item{pi_model_type}{Type of propensity score model. Options are 'glm' (GLM).
or 'rf' (random forest)}

\item{pi_model_seed}{Random seed for random forest propensity score model.}

\item{f_lasso}{Formula for GLM propensity score model.}

\item{defs_length_base}{Base length for output of get_defs_analysis.}
}
\value{
List with the following components:
\itemize{
\item ipw: Matrix of null unfairness metrics, one row for each replication.
}
}
\description{
Simulate null distributions.
}
