% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_fairness.R
\name{bs_rescaled_analysis}
\alias{bs_rescaled_analysis}
\title{Rescaled, stratified bootstrap for unfairness metrics}
\usage{
bs_rescaled_analysis(
  data,
  cutoff,
  B,
  m_factor = 0.75,
  pi_model_type,
  pi_model_seed = NULL,
  f_lasso = NULL,
  xvars,
  defs_length_base
)
}
\arguments{
\item{data}{Dataframe or tibble. Must include at least A1, A2, Y (binary 0/1),
D (binary 0/1), covariates used to train propensity score model,
S_prob (continuous risk probability).}

\item{cutoff}{Classification cutoff. Must be a single numeric value.}

\item{B}{Number of bootstrap replications.}

\item{m_factor}{Fractional power for calculating resample size (default 0.75).}

\item{pi_model_type}{Type of propensity score model. Options are 'glm' (GLM).
or 'rf' (random forest)}

\item{pi_model_seed}{Random seed for random forest propensity score model.}

\item{f_lasso}{Formula for GLM propensity score model.}

\item{xvars}{Character vector of covariates used in propensity score model.}

\item{defs_length_base}{Base length for output of get_defs_analysis.}
}
\value{
A matrix of definitions, one row for each of the B bootstrap replications.
}
\description{
Rescaled, stratified bootstrap for unfairness metrics
}
