% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_format.R
\name{ci_norm}
\alias{ci_norm}
\title{Calculate normal approximation confidence interval.}
\usage{
ci_norm(bs_rescaled, est_named, parameter, sampsize, alpha)
}
\arguments{
\item{bs_rescaled}{Rescaled bootstrap estimate table (result of get_bs_rescaled).}

\item{est_named}{Named 1-row dataframe of raw metric estimates (result of names_defs).}

\item{parameter}{Column name of the metric being estimated (string).}

\item{sampsize}{Sample size of estimation data set.}

\item{alpha}{Size of (1-alpha) confidence interval.}
}
\value{
Dataframe with point estimate, SE, variance, 1-alpha confidence interval.
}
\description{
Calculate normal approximation confidence interval.
}
