% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_format.R
\name{ci_trunc}
\alias{ci_trunc}
\title{Truncate confidence interval.}
\usage{
ci_trunc(ci_result, type)
}
\arguments{
\item{ci_result}{Confidence interval dataframe (result of ci_norm or ci_tint).}

\item{type}{Type of confidence interval ('norm' or 'tint')}
}
\value{
Confidence interval dataframe with truncated interval.
}
\description{
Truncate confidence interval.
}
