% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_format.R
\name{defs_names}
\alias{defs_names}
\title{Get vector of names for output of 'get_defs_analysis' function.}
\usage{
defs_names(A1_length, A2_length)
}
\arguments{
\item{A1_length}{Length of 'A1' protected characteristic variable.}

\item{A2_length}{Length of 'A2' protected characteristic variable.}
}
\value{
Vector of names.
}
\description{
Get vector of names for output of 'get_defs_analysis' function.
}
