% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ex_data_external}
\alias{ex_data_external}
\title{Simulated 'external data' for demonstration of small subgroup unfairness metrics with data borrowing.}
\format{
\subsection{\code{ex_data_external}}{

A data frame with 5,000 rows and 25 columns:
\describe{
\item{A1, A2}{Protected characteristics}
\item{X_pa.1, ..., X_pa.5}{Covariates for external P(A=a) model (numeric)}
...
}
}
}
\source{
Generated via simulation.
}
\usage{
ex_data_external
}
\description{
Simulated 'external data' for demonstration of small subgroup unfairness metrics with data borrowing.
}
\keyword{datasets}
