% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_format.R
\name{get_bs_rescaled}
\alias{get_bs_rescaled}
\title{Create rescaled bootstrap estimate table from results of bs_rescaled_analysis.}
\usage{
get_bs_rescaled(bs_table, defs, sampsize, A1_length, A2_length, m_factor)
}
\arguments{
\item{bs_table}{Dataframe of bootstrap results.}

\item{defs}{List of definitions, output by 'get_defs_analysis' function.}

\item{sampsize}{Sample size of estimation data set.}

\item{A1_length}{Length of 'A1' protected characteristic variable.}

\item{A2_length}{Length of 'A2' protected characteristic variable.}

\item{m_factor}{Fractional power for calculating resample size (default 0.75).}
}
\value{
Dataframe of rescaled bootstrap estimates.
}
\description{
Create rescaled bootstrap estimate table from results of bs_rescaled_analysis.
}
