% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_fairness.R
\name{get_defs_analysis}
\alias{get_defs_analysis}
\title{Get estimated unfairness metrics}
\usage{
get_defs_analysis(data, cutoff)
}
\arguments{
\item{data}{Dataframe or tibble. Must include at least A1, A2, Y (binary 0/1),
D (binary 0/1), Y0 (estimated), pi (estimated).}

\item{cutoff}{Classification cutoff. Must be a single numeric value.}
}
\value{
A list with the following components:
\itemize{
\item defs: vector of unfairness definitions with (12+4)nlevels(A1)nlevels(A2) items
}
}
\description{
'get_defs_analysis' returns a list of unfairness definitions.
}
\details{
Note: Indices and/or order of definitions in output of this function are
used in formatting functions.
}
