% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_fairness.R
\name{get_est_analysis}
\alias{get_est_analysis}
\title{Get estimated propensity score and $Y^0$ weighted estimator}
\usage{
get_est_analysis(
  data,
  cutoff,
  pi_model = NULL,
  pi_model_type,
  pi_model_seed = NULL,
  f_lasso = NULL,
  xvars
)
}
\arguments{
\item{data}{Dataframe or tibble. Must include at least A1, A2, Y (binary 0/1),
D (binary 0/1), covariates used to train propensity score model,
S_prob (continuous risk probability).}

\item{cutoff}{Classification cutoff. Must be a single numeric value.}

\item{pi_model}{Propensity score model object.}

\item{pi_model_type}{Type of propensity score model. Options are 'glm' (GLM).
or 'rf' (random forest)}

\item{pi_model_seed}{Random seed for random forest propensity score model.}

\item{f_lasso}{Formula for GLM propensity score model.}

\item{xvars}{Character vector of covariates used in propensity score model.}
}
\value{
List with the following components:
\itemize{
\item Y0est: Vector of inverse probability weighted estimates of $Y^0$.
\item pi: Vector of estimated propensity scores.
}
}
\description{
Get estimated propensity score and $Y^0$ weighted estimator
}
