% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_smallsubgroups.R
\name{get_pa_ext_small}
\alias{get_pa_ext_small}
\title{Fit external data P(A=a) model.}
\usage{
get_pa_ext_small(data_external, pa_xvars_ext, fit_method_ext, maxit = 1000)
}
\arguments{
\item{data_external}{External data set. Must contain at least A1, A2, and a subset of covariates used to train P(A=a) model.}

\item{pa_xvars_ext}{Character vector of covariates used in external P(A=a) model.}

\item{fit_method_ext}{Type of external data model for P(A=a): multinomial ('multinomial') or neural network ('neural_net')}

\item{maxit}{Maximum number of iterations for neural_net model. Default 1000.}
}
\value{
Model for P(A=a|X) fit on external data.
}
\description{
Fit external data P(A=a) model.
}
