% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_format.R
\name{names_defs}
\alias{names_defs}
\title{Assign column names to output of 'get_defs_analysis' or other list result object.}
\usage{
names_defs(defs, A1_length, A2_length)
}
\arguments{
\item{defs}{List of definitions, output by 'get_defs_analysis' function.}

\item{A1_length}{Length of 'A1' protected characteristic variable.}

\item{A2_length}{Length of 'A2' protected characteristic variable.}
}
\value{
Dataframe of unfairness metrics with named columns.
}
\description{
Assign column names to output of 'get_defs_analysis' or other list result object.
}
