% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_format.R
\name{names_df}
\alias{names_df}
\title{Assign column names to bootstrap results or other dataframe result object.}
\usage{
names_df(bs_table, A1_length, A2_length)
}
\arguments{
\item{bs_table}{Dataframe of bootstrap results.}

\item{A1_length}{Length of 'A1' protected characteristic variable.}

\item{A2_length}{Length of 'A2' protected characteristic variable.}
}
\value{
Dataframe of bootstrap results with named columns.
}
\description{
Assign column names to bootstrap results or other dataframe result object.
}
