% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_format.R
\name{select_coef}
\alias{select_coef}
\title{Extract names of variables with non-zero coefficients after cv.glmnet.}
\usage{
select_coef(model, lambda = NULL, xvars, outcome)
}
\arguments{
\item{model}{cv.glmnet model object.}

\item{lambda}{Optional lambda value for selecting coefficients. Defaults to lambda.1se.}

\item{xvars}{Variable names for all predictors in cv.glmnet object.}

\item{outcome}{Outcome variable name (for formula)}
}
\value{
A list with the following components:
\itemize{
\item coef: Dataframe with non-zero coefficients
\item vars_names: Names of variables with non-zero coefficients
\item formula: Formula with selected variables
}
}
\description{
Extract names of variables with non-zero coefficients after cv.glmnet.
}
