% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_fairness.R
\name{analysis_estimation}
\alias{analysis_estimation}
\title{Main function: Estimation of nuisance parameters and unfairness metrics.}
\usage{
analysis_estimation(
  data,
  cutoff,
  estimator_type = c("standard", "small_internal", "small_borrow"),
  gen_null = c(T, F),
  R_null = 500,
  bootstrap = c("none", "rescaled"),
  B = 500,
  m_factor = 0.75,
  pi_model = NULL,
  pi_model_type = c("glm", "rf"),
  pi_model_seed = NULL,
  pi_xvars,
  outcome_model_type = c(NULL, "glm", "rf", "neural_net"),
  outcome_xvars = NULL,
  fit_method_int = c(NULL, "multinomial", "neural_net"),
  nfolds = 5,
  pa_xvars_int = NULL,
  data_external = NULL,
  fit_method_ext = c(NULL, "multinomial", "neural_net"),
  pa_xvars_ext = NULL,
  borrow_metric = c(NULL, "auc", "brier"),
  pa_model_ext = NULL
)
}
\arguments{
\item{data}{Dataframe or tibble. Must include at least A1, A2, Y (binary 0/1),
D (binary 0/1), covariates used to train propensity score model,
S_prob (probability).}

\item{cutoff}{Classification cutoff. Must be a single numeric value.}

\item{estimator_type}{Type of estimation: standard ('standard'), small subgroup internal-only ('small_internal'),
or small subgroup with data borrowing ('small_borrow').}

\item{gen_null}{T/F: generate null distributions.}

\item{R_null}{Number of replications for permutation null distribution (default 500).}

\item{bootstrap}{Obtain bootstrap estimates using rescaled method ('rescaled') or
no bootstrap ('none', the default).}

\item{B}{Number of bootstrap resamples (default 500).}

\item{m_factor}{Fractional power for calculating resample size (default 0.75).}

\item{pi_model}{Pre-fit propensity score model.}

\item{pi_model_type}{Type of propensity score model: logistic regression ('glm') or random forest ('rf')}

\item{pi_model_seed}{Numeric vector of random seeds for random forest propensity score model. Required if pi_model_type is 'rf'.
Must be length 2 if pi_model is not specified and length 1 if pi_model is specified.
First element is set prior to model fitting; second element (or only element if pi_model specified) is set prior to prediction.}

\item{pi_xvars}{Character vector of covariates used in propensity score model.}

\item{outcome_model_type}{Type of outcome model: logistic regression ('glm'), random forest ('rf'),
or neural network ('neural_net')}

\item{outcome_xvars}{Character vector of covariates used in outcome models.}

\item{fit_method_int}{Type of internal data model for P(A=a): multinomial ('multinomial') or neural network ('neural_net')}

\item{nfolds}{Number of folds for cross-fitting (default 5).}

\item{pa_xvars_int}{Character vector of covariates used in internal P(A=a) model.}

\item{data_external}{External data set. Must contain at least A1, A2, and a subset of covariates used to train P(A=a) model.}

\item{fit_method_ext}{Type of external data model for P(A=a): multinomial ('multinomial') or neural network ('neural_net')}

\item{pa_xvars_ext}{Character vector of covariates used in external P(A=a) model.}

\item{borrow_metric}{Metric for data borrowing: AUC ('auc') or Brier score ('brier')}

\item{pa_model_ext}{Pre-fit external P(A=a) model. Can be either multinomial or neural network model produced by the package nnet.
Model type must match fit_method_ext.}
}
\value{
List with the following components:
\itemize{
\item defs: estimated definitions
\item estY0: mean estimated Y0 for each protected group
\item table_null: null distribution table (if gen_null = T)
\item boot_out: bootstrap estimates for metrics and error rates (if bootstrap = 'rescaled')
\item est.choice: Input data frame with Y0 estimates added
}
}
\description{
Main function: Estimation of nuisance parameters and unfairness metrics.
}
