% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_plots.R
\name{get_plots}
\alias{get_plots}
\title{Plot results of 'analysis_estimation'.}
\usage{
get_plots(
  results,
  sampsize,
  alpha,
  m_factor,
  A1_length,
  A2_length,
  plot_labels,
  plot_values,
  plot_colors,
  delta_uval,
  uval_labels
)
}
\arguments{
\item{results}{Results of 'analysis_estimation', including null distribution and bootstrap.
Must be a list with at least the components: defs, table_null, boot_out.}

\item{sampsize}{Sample size of estimation data set.}

\item{alpha}{Size of confidence interval.}

\item{m_factor}{Fractional power for calculating resample size.}

\item{A1_length}{Number of levels of A1.}

\item{A2_length}{Number of levels of A2.}

\item{plot_labels}{Vector of (shortened) protected group names for labeling counterfactual error
rate plots. Must have length(A1)*length(A2) components.}

\item{plot_values}{Vector of numeric ggplot2 shape scale values
for counterfactual error rate plots. Must have length(A1)*length(A2) components.}

\item{plot_colors}{Vector of colors (string name or hex) for counterfactual error
rate plots. Must have length(A1)*length(A2) components.}

\item{delta_uval}{Threshold for u-value test. Must be between 0 and 1.}

\item{uval_labels}{Vector of x and y coordinates for labels on u-value plots. Must have length
12.}
}
\value{
List of plots with the following components:
\itemize{
\item cfpr: Counterfactual false positive rates by group.
\item cfnr: Counterfactual false negative rates by group.
\item metrics_pos: Positive unfairness metrics.
\item metrics_neg: Negative unfairness metrics.
\item null_dist: Null distribution plots with estimated metrics and confidence intervals.
}
}
\description{
Plot results of 'analysis_estimation'.
}
